[refactor] Admin에 관련된 패키지를 묶기
