+++ src/main/java/team7/inplace/admin/AdminPageController.java
@@ -9,9 +9,9 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import team7.inplace.admin.persistence.BannerRepository;
-import team7.inplace.global.exception.ErrorLog;
-import team7.inplace.global.exception.ErrorLogRepository;
+import team7.inplace.admin.banner.persistence.BannerRepository;
+import team7.inplace.admin.error.ErrorLog;
+import team7.inplace.admin.error.ErrorLogRepository;
 import team7.inplace.global.kakao.config.KakaoApiProperties;
 import team7.inplace.video.domain.Video;
 import team7.inplace.video.persistence.VideoRepository;
