+++ src/main/java/team7/inplace/admin/crawling/client/dto/PlaceNode.java
@@ -1,0 +1,12 @@
+package team7.inplace.admin.crawling.client.dto;
+
+import com.fasterxml.jackson.databind.JsonNode;
+
+public record PlaceNode(
+        JsonNode locationNode,
+        JsonNode placeNode
+) {
+    public static PlaceNode of(JsonNode locationNode, JsonNode placeNode) {
+        return new PlaceNode(locationNode, placeNode);
+    }
+}
