+++ src/main/java/team7/inplace/admin/error/ErrorLogRepository.java
@@ -1,0 +1,8 @@
+package team7.inplace.admin.error;
+
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ErrorLogRepository extends JpaRepository<ErrorLog, Long> {
+    List<ErrorLog> findByIsResolvedFalse();
+}
