+++ src/main/java/team7/inplace/video/application/VideoFacade.java
@@ -1,13 +1,10 @@
 package team7.inplace.video.application;
 
 import java.util.List;
-import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.transaction.annotation.Transactional;
-import team7.inplace.crawling.application.YoutubeCrawlingService;
+import team7.inplace.admin.crawling.application.YoutubeCrawlingService;
 import team7.inplace.global.annotation.Facade;
 import team7.inplace.global.exception.InplaceException;
 import team7.inplace.global.exception.code.AuthorizationErrorCode;
