+++ src/test/java/team7/inplace/crawling/application/AddressUtilTest.java
@@ -4,6 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import team7.inplace.admin.crawling.application.AddressUtil;
 
 class AddressUtilTest {
