+++ src/test/java/team7/inplace/crawling/client/YoutubeClientTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
+import team7.inplace.admin.crawling.client.YoutubeClient;
 
 @SpringBootTest
 @ActiveProfiles("test")
